requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
python-dotenv==1.0.1
groq==0.4.0
tenacity==8.2.3
python-dateutil==2.8.2
requests==2.31.0